requirements.txt
Flask==3.1.2
google-generativeai==0.8.5
